requirements.txt
google_generativeai
pyaudio
faster-whisper
SpeechRecognition
openai
python-dotenv